pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nexari"
version = "0.1.0"
description = "From intent to deployed model. Autonomous ML pipeline agent."
readme = "README.md"
requires-python = ">=3.10"
license = "Apache-2.0"
authors = [{ name = "J.D. Aguilera" }]
keywords = ["machine-learning", "mlops", "autonomous-agent", "huggingface", "bedrock"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "boto3>=1.34",
    "huggingface-hub>=0.23",
    "datasets>=2.19",
    "transformers>=4.40",
    "fastapi>=0.111",
    "uvicorn>=0.29",
    "pydantic>=2.7",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
ollama = ["ollama>=0.2"]
dev = ["pytest>=8", "pytest-asyncio>=0.23", "ruff>=0.4", "mypy>=1.10"]

[project.scripts]
nexari = "nexari.cli:main"

[project.urls]
Homepage = "https://github.com/jdavidaguil/nexari"
Issues = "https://github.com/jdavidaguil/nexari/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["nexari*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.10"
strict = false